+++ ligo/skymap/plot/tests/test_plot.py
@@ -116,7 +116,12 @@ def test_zoom_axes(rcparams):
     ax.scalebar((0.1, 0.1), 30 * u.arcmin)
     ax.grid()
     for key in ['ra', 'dec']:
-        ax.coords[key].set_auto_axislabel(False)
+        try:
+            ax.coords[key].set_auto_axislabel(False)
+        except AttributeError:
+            # FIXME: CoordinateHelper.set_auto_axislabel was added in
+            # astropy-4.0, but we currently support astropy >= 3.1.
+            pass
     return fig